// src/core/types/UserInfos.ts
 
 

 export type userResponse = {

    login: string;
    public_repos: number;
    location: string;
    created_at: Date;
    avatar_url:string;
    followers:number;
    following: number;
    html_url: string;
    company:string;
    blog:string;
}




export type UserInfos = {

    login: string;
    public_repos: number;
    location: string;
    created_at: Date;
    avatar_url:string;
    following: number;
    html_url: string;
    followers:number;
    company:string;
    blog:string;
 }